+++ src/app.module.ts
@@ -7,8 +7,8 @@ import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { BoardModule } from './board/board.module';
 import { CardModule } from './card/card.module';
-import { CheckItemModule } from './check_item/check_item.module';
-import { CheckListModule } from './check_list/check_list.module';
+import { CheckItemModule } from './checkitem/checkitem.module';
+import { CheckListModule } from './checklist/checklist.module';
 import { CommentModule } from './comment/comment.module';
 import { ListModule } from './list/list.module';
 import { AppController } from './app.controller';
