--- src/check_list/dto/create-check_list.dto.ts
@@ -1,1 +1,0 @@
-export class CreateCheckListDto {}
