--- src/check_list/dto/update-check_list.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateCheckListDto } from './create-check_list.dto';
-
-export class UpdateCheckListDto extends PartialType(CreateCheckListDto) {}
