+++ src/checkitem/dto/update-checkitem.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/swagger';
-import { CreateCheckItemDto } from './create-check_item.dto';
+import { CreateCheckItemDto } from './create-checkitem.dto';
 
 export class UpdateCheckItemDto extends PartialType(CreateCheckItemDto) {}
