+++ src/checklist/dto/update-checklist.dto.ts
@@ -1,0 +1,6 @@
+import { PickType } from '@nestjs/swagger';
+import { CreateCheckListDto } from './create-checklist.dto';
+
+export class UpdateCheckListDto extends PickType(CreateCheckListDto, [
+  'title',
+]) {}
